+++ src/Pages/CartPage/CartPage.tsx
@@ -8,6 +8,7 @@ import CheckCircleBtn from "../../components/common/CheckBtn/CheckCircleBtn";
 import { CartItemType } from "../../types/Cart.type";
 import useFetchCartItems from "../../hooks/queries/useFetchCartItems";
 import { setPaymentAmount } from "../../features/paymentAmountSlice";
+import Loading from "../../components/common/Loading/Loading";
 
 function CartPage() {
   const dispatch = useAppDispatch();
@@ -109,24 +110,26 @@ function CartPage() {
         <li>수량</li>
         <li>상품금액</li>
       </S.MenuUl>
-      {cartItems.map((cartItem: CartItemType, index: number) => (
-        <CartItem
-          key={cartItem.cart_item_id}
-          cartItemId={cartItem.cart_item_id}
-          quantity={cartItem.quantity}
-          cartItem={cartItem.productDetail}
-          isChecked={isCheckedArray[index]}
-          onToggle={() => handleToggleCheckbox(index)}
-        />
-      ))}
-      {cartItems.length === 0 ? (
-        <S.NoItemBox>
-          <p>장바구니에 담긴 상품이 없습니다.</p>
-          <small>원하는 상품을 장바구니에 담아보세요!</small>
-        </S.NoItemBox>
-      ) : (
-        <TotalPrice />
-      )}
+      <Suspense fallback={<Loading />}>
+        {cartItems.map((cartItem: CartItemType, index: number) => (
+          <CartItem
+            key={cartItem.cart_item_id}
+            cartItemId={cartItem.cart_item_id}
+            quantity={cartItem.quantity}
+            cartItem={cartItem.productDetail}
+            isChecked={isCheckedArray[index]}
+            onToggle={() => handleToggleCheckbox(index)}
+          />
+        ))}
+        {cartItems.length === 0 ? (
+          <S.NoItemBox>
+            <p>장바구니에 담긴 상품이 없습니다.</p>
+            <small>원하는 상품을 장바구니에 담아보세요!</small>
+          </S.NoItemBox>
+        ) : (
+          <TotalPrice />
+        )}
+      </Suspense>
     </S.CartPageLayout>
   );
 }
